+++ libraries/template-haskell/Language/Haskell/TH.hs
@@ -20,7 +20,6 @@ module Language.Haskell.TH(
         -- *** Reify
         reify,            -- :: Name -> Q Info
         reifyModule,
-        thisModule,
         Info(..), ModuleInfo(..),
         InstanceDec,
         ParentName,
@@ -84,89 +83,7 @@ module Language.Haskell.TH(
         FamilyResultSig(..), Syntax.InjectivityAnn(..), PatSynType,
 
     -- * Library functions
-    -- ** Abbreviations
-        InfoQ, ExpQ, DecQ, DecsQ, ConQ, TypeQ, TyLitQ, CxtQ, PredQ,
-        DerivClauseQ, MatchQ, ClauseQ, BodyQ, GuardQ, StmtQ, RangeQ,
-        SourceStrictnessQ, SourceUnpackednessQ, BangTypeQ, VarBangTypeQ,
-        StrictTypeQ, VarStrictTypeQ, PatQ, FieldPatQ, RuleBndrQ, TySynEqnQ,
-        PatSynDirQ, PatSynArgsQ,
-
-    -- ** Constructors lifted to 'Q'
-    -- *** Literals
-        intPrimL, wordPrimL, floatPrimL, doublePrimL, integerL, rationalL,
-        charL, stringL, stringPrimL, charPrimL,
-    -- *** Patterns
-        litP, varP, tupP, unboxedSumP, conP, uInfixP, parensP, infixP,
-        tildeP, bangP, asP, wildP, recP,
-        listP, sigP, viewP,
-        fieldPat,
-
-    -- *** Pattern Guards
-        normalB, guardedB, normalG, normalGE, patG, patGE, match, clause,
-
-    -- *** Expressions
-        dyn, varE, conE, litE, appE, appTypeE, uInfixE, parensE, staticE,
-        infixE, infixApp, sectionL, sectionR,
-        lamE, lam1E, lamCaseE, tupE, unboxedSumE, condE, multiIfE, letE, caseE,
-        appsE, listE, sigE, recConE, recUpdE, stringE, fieldExp,
-    -- **** Ranges
-    fromE, fromThenE, fromToE, fromThenToE,
-
-    -- ***** Ranges with more indirection
-    arithSeqE,
-    fromR, fromThenR, fromToR, fromThenToR,
-    -- **** Statements
-    doE, compE,
-    bindS, letS, noBindS, parS,
-
-    -- *** Types
-        forallT, varT, conT, appT, arrowT, infixT, uInfixT, parensT, equalityT,
-        listT, tupleT, unboxedSumT, sigT, litT, promotedT, promotedTupleT,
-        promotedNilT, promotedConsT,
-    -- **** Type literals
-    numTyLit, strTyLit,
-    -- **** Strictness
-    noSourceUnpackedness, sourceNoUnpack, sourceUnpack,
-    noSourceStrictness, sourceLazy, sourceStrict,
-    isStrict, notStrict, unpacked,
-    bang, bangType, varBangType, strictType, varStrictType,
-    -- **** Class Contexts
-    cxt, classP, equalP,
-    -- **** Constructors
-    normalC, recC, infixC, forallC, gadtC, recGadtC,
-
-    -- *** Kinds
-    varK, conK, tupleK, arrowK, listK, appK, starK, constraintK,
-
-    -- *** Roles
-    nominalR, representationalR, phantomR, inferR,
-
-    -- *** Top Level Declarations
-    -- **** Data
-    valD, funD, tySynD, dataD, newtypeD,
-    derivClause, DerivClause(..), DerivStrategy(..),
-    -- **** Class
-    classD, instanceD, instanceWithOverlapD, Overlap(..),
-    sigD, standaloneDerivD, standaloneDerivWithStrategyD, defaultSigD,
-
-    -- **** Role annotations
-    roleAnnotD,
-    -- **** Type Family / Data Family
-    dataFamilyD, openTypeFamilyD, closedTypeFamilyD, dataInstD,
-    familyNoKindD, familyKindD, closedTypeFamilyNoKindD, closedTypeFamilyKindD,
-    newtypeInstD, tySynInstD,
-    typeFam, dataFam, tySynEqn, injectivityAnn, noSig, kindSig, tyVarSig,
-    -- **** Foreign Function Interface (FFI)
-    cCall, stdCall, cApi, prim, javaScript,
-    unsafe, safe, forImpD,
-    -- **** Pragmas
-    ruleVar, typedRuleVar,
-    pragInlD, pragSpecD, pragSpecInlD, pragSpecInstD, pragRuleD, pragAnnD,
-    pragLineD,
-
-    -- **** Pattern Synonyms
-    patSynD, patSynSigD, unidir, implBidir, explBidir, prefixPatSyn,
-    infixPatSyn, recordPatSyn,
+    module Language.Haskell.TH.Lib,
 
     -- * Pretty-printer
     Ppr(..), pprint, pprExp, pprLit, pprPat, pprParendType